+++ muffnn/autoencoder/tests/test_autoencoder.py
@@ -1,23 +1,17 @@
 """
 Tests for the Autoencoder.
 """
-from __future__ import print_function
-from __future__ import division
 
 import logging
 import pprint
 from io import BytesIO
 import pickle
 import sys
-try:
-    from unittest import mock
-except ImportError:
-    mock = None
+from unittest import mock
 
 import pytest
 import numpy as np
 import scipy.sparse as sp
-import six
 import tensorflow as tf
 from sklearn.preprocessing import MinMaxScaler, OneHotEncoder
 from sklearn.datasets import load_iris
@@ -26,11 +20,6 @@ from sklearn.utils.testing import (assert_array_almost_equal,
                                    assert_almost_equal)
 
 from muffnn import Autoencoder
-
-if six.PY2:
-    from mock import patch
-else:
-    from unittest.mock import patch
 
 
 _LOGGER = logging.getLogger(__name__)
@@ -112,7 +101,7 @@ class AutoencoderManyEpochs(Autoencoder):
 # replicate outside of travis, but I was able to get the test to fail locally
 # by changing atol in sklearn.utils.check_methods_subset_invariance from 1e-7
 # to 1e-10. This simply skips that part of check_estimator.
-@patch('sklearn.utils.estimator_checks.check_methods_subset_invariance')
+@mock.patch('sklearn.utils.estimator_checks.check_methods_subset_invariance')
 def test_check_estimator(mock_check_methods_subset_invariance):
     """Check adherence to Estimator API."""
     if sys.version_info.major == 3 and sys.version_info.minor == 7: