+++ muffnn/fm/tests/test_fm_regressor.py
@@ -4,16 +4,11 @@ Tests for the FM Regressor
 based in part on sklearn's logistic tests:
 https://github.com/scikit-learn/scikit-learn/blob/master/sklearn/linear_model/tests/test_logistic.py
 """
-from __future__ import print_function
-from __future__ import division
 
 from io import BytesIO
 import pickle
 import sys
-try:
-    from unittest import mock
-except ImportError:
-    mock = None
+from unittest import mock
 
 import numpy as np
 import pytest